type modification

+++ Robot/src/main/java/frc/robot/Config.java
@@ -4,12 +4,12 @@ public class Config {
 
     // CAN BUS port
     // Swerve
-    public static double front_left_drive  = 1;
-    public static double front_left_steer  = 5;
-    public static double front_right_drive = 2;
-    public static double front_right_steer = 6;
-    public static double back_left_drive   = 10;
-    public static double back_left_steer   = 7;
-    public static double back_right_drive  = 11;
-    public static double back_right_steer  = 8;
+    public static int front_left_drive  = 1;
+    public static int front_left_steer  = 5;
+    public static int front_right_drive = 2;
+    public static int front_right_steer = 6;
+    public static int back_left_drive   = 10;
+    public static int back_left_steer   = 7;
+    public static int back_right_drive  = 11;
+    public static int back_right_steer  = 8;
 }
